New error output + bug fixes + leaks patched.
Commands that couldn't be parsed correctly (currently only happens if a
command starts with a ;), have some nice output printed out. This is
technically a bug fix too as this would crash the program...?

Switched cmd in main to a pointer. Fixed memory leaks that came with
that + memory leaks that already existed (because rip continue...).