Added 'if' statements
Turns out things like if command; then if, or while command; do if, if
command; then while, and while command; do while, do not work.

Which makes sense, because the logic that waits for a complete while
loop only runs if the first token of a command is while, but in the last
2 examples, the first token is then, and do respectively.

The solution will likely be, to move that part that waits for finished
while/if blocks, to its own function, so I can easily call it when I
encounter 'do' or 'then' followed by 'while' or 'if'.

Still, that is somewhat low priority, since the easiest workaround is to
just press enter and put the next while/if on its own line! (Or just add
a semicolon, although I may have forgotten to test for empty commands so
'do ;' and 'then ;' might cause issues...)