Lowered memory usage when parsing commands.
Instead of the shotgun approach of allocating one argument for every
space in the buffer, mash properly pre-parses the input, to determine
how many arguments there are, then parses the arguments.

While this may decrease performance, it may also make no difference (or
even improve performance) because we technically were already parsing
the string (finding every space).

I say it could improve performance, because with the current setup, we
don't have to parse the entire string. If there is a semicolon,
everything after the semicolon will be ignored, thus, parsing quicker.