Pipes! But...
"Minor" regression with file IO redirection. It still works, but not
with while/if. Or at least not entirely?

while read line; do
  echo "Got: $line"
done < file

The above works correctly, but in the code below, cat never outputs
anything, and all of the contents of the file are passed to `read`...

while read line; do
  echo "Got: $line"
  cat
done < file