"Fixed" IO redirection + fixed threads + more
- Replaced several instances of fprintf with more appropriate functions
  like fwrite, fputs, fputc.
- Simplified logic for commandFree since commands no longer recursively
  point to each other (like they did with while loops), so less checks
  are needed for safety.
- Fixed threads. As in, I removed the debug messages, but also fixed
  some really stupid typos.
- Updated various error messages.
- "Fixed" IO redirection regressions, more below.

I have found what is either a bug in glibc (though I discovered the same
behavior in bionic libc), or something that is expected but very stupid
behavior.

Here is a small sample demonstrating the issue, which requires a text
file named 'text' with at least 2 lines.

(include stdio, stdlib, and unistd)

int main(void) {
    FILE *file = fopen("text", "r");
    int fd = fileno(file);
    off_t offset = lseek(fd, 0, SEEK_CUR);

    char *line = NULL;
    size_t size = 0;
    ssize_t bytes_read = getline(&line, &size, file);
    printf(
        "getline() returned %zi, and read: <%s>\n",
        bytes_read,
        line);
    free(line);

    off_t after = lseek(fd, 0, SEEK_CUR);
    printf(
        "offset started at %ld, and moved to %ld - EOF is %ld\n",
        offset,
        after,
        lseek(fd, 0, SEEK_END));
    lseek(fd, after, SEEK_SET);

    char remaining[512] = {};
    bytes_read = read(fd, remaining, 511);
    printf(
        "read() returned %zi, and read: <%s>\n",
        bytes_read,
        remaining);
}

To explain what happens without forcing you to compile the code however,
getline reads in one line from the file and then I print it. But the
call to read returns 0 and doesn't read anything. Why? Because after
calling getline, the file descriptor's offset is at the end of the file!

That is why the commit that added pipes had this code break:

while read line; do
  echo "got: $line"
  cat
done < file

It would print every line with echo, because cat didn't get any data.
Why not? Well if we look at my above sample, we see why. Cat is calling
fileno on the file passed to the while loop, but it's at the end of the
file, so cat always reads 0 bytes, allowing read to continue grabbing
data (since read is a builtin command that calls getline).

This issue happens with getline, fread, fgets - doesn't matter. I have
found nothing about this online, nor any mentions in bug trackers. So
for now my workaround was some gross code that involves manually syncing
the positions of the FILE and the fd. It's even more cursed because
fseek and lseek technically use different types (long vs off_t).