Added $# and shift.
$# expands to the number of arguments in the given program/script call.
I don't actually know if $# will ever be 0 in an interactive shell, but
when using mash to run scripts, that's where you'd likely use this.

If you have a script 'test.mash', and you call it like this:

mash test.mash hello world

then $# expands to 2, $1 expands to 'hello', and $2 expands to 'world'.
If you then run `shift`, $# will expand to 1, $1 will expand to 'world',
and $2 will no longer expand to anything.