Remove Suffix Tree files.
Sad day, I was kinda proud of those even if they were a little whack,
but I never really utilized them much and I knew it was better to not
force such a rigid structure. Just look at the difference in arguments
passed to all the various builtins, to allow them to do what they need.