Makefile fix for some compilers.
On my system, the order of the -l library links doesn't seem to matter,
but I discovered on some systems having them as the first arguments
caused the linking process to fail, and moving them to the end fixed the
issue.

Also updated the README with some mostly useless information.